project/binary-python/requirements.txt
pycryptodome==3.19.0
pyserial==3.5
